+++ api/prisma/seed-dev.ts
@@ -47,7 +47,7 @@ export const devSeeding = async (
   const jurisdiction = await prismaClient.jurisdictions.create({
     data: {
       ...jurisdictionFactory(jurisdictionName),
-      allowSingleUseCodeLogin: true,
+      allowSingleUseCodeLogin: false,
     },
   });
   await prismaClient.userAccounts.create({
