+++ api/src/controllers/auth.controller.ts
@@ -64,7 +64,10 @@ export class AuthController {
     @Request() req: ExpressRequest,
     @Response({ passthrough: true }) res: ExpressResponse,
   ): Promise<SuccessDTO> {
-    return await this.authService.setCredentials(res, mapTo(User, req['user']));
+    return await this.authService.confirmAndSetCredentials(
+      mapTo(User, req['user']),
+      res,
+    );
   }
 
   @Get('logout')
@@ -88,19 +91,6 @@ export class AuthController {
     @Body() dto: RequestMfaCode,
   ): Promise<RequestMfaCodeResponse> {
     return await this.authService.requestMfaCode(dto);
-  }
-
-  @Post('request-single-use-code')
-  @ApiOperation({
-    summary: 'Request single use code',
-    operationId: 'requestSingleUseCode',
-  })
-  @ApiOkResponse({ type: SuccessDTO })
-  async requestSingleUseCode(
-    @Request() req: ExpressRequest,
-    @Body() dto: RequestSingleUseCode,
-  ): Promise<SuccessDTO> {
-    return await this.authService.requestSingleUseCode(dto, req);
   }
 
   @Get('requestNewToken')
