+++ api/src/services/user.service.ts
@@ -11,6 +11,8 @@ import dayjs from 'dayjs';
 import advancedFormat from 'dayjs/plugin/advancedFormat';
 import crypto from 'crypto';
 import { verify, sign } from 'jsonwebtoken';
+import { Request } from 'express';
+
 import { PrismaService } from './prisma.service';
 import { User } from '../dtos/users/user.dto';
 import { mapTo } from '../utilities/mapTo';
@@ -37,6 +39,8 @@ import { permissionActions } from '../enums/permissions/permission-actions-enum'
 import { buildWhereClause } from '../utilities/build-user-where';
 import { getPublicEmailURL } from '../utilities/get-public-email-url';
 import { UserRole } from '../dtos/users/user-role.dto';
+import { RequestSingleUseCode } from '../dtos/single-use-code/request-single-use-code.dto';
+import { generateSingleUseCode } from '../utilities/generate-single-use-code';
 
 /*
   this is the service for users
@@ -479,9 +483,11 @@ export class UserService {
     dto: UserCreate | UserInvite,
     forPartners: boolean,
     sendWelcomeEmail = false,
-    requestingUser: User,
-    jurisdictionName?: string,
+    req: Request,
   ): Promise<User> {
+    const requestingUser = mapTo(User, req['user']);
+    const jurisdictionName = (req.headers['jurisdictionname'] as string) || '';
+
     if (
       this.containsInvalidCharacters(dto.firstName) ||
       this.containsInvalidCharacters(dto.lastName)
@@ -646,16 +652,27 @@ export class UserService {
 
     // Public user that needs email
     if (!forPartners && sendWelcomeEmail) {
-      const confirmationUrl = this.getPublicConfirmationUrl(
-        dto.appUrl,
-        confirmationToken,
-      );
-      this.emailService.welcome(
-        jurisdictionName,
-        mapTo(User, newUser),
-        dto.appUrl,
-        confirmationUrl,
-      );
+      const fullJurisdiction = await this.prisma.jurisdictions.findFirst({
+        where: {
+          name: jurisdictionName as string,
+        },
+      });
+
+      if (fullJurisdiction?.allowSingleUseCodeLogin) {
+        this.requestSingleUseCode(dto, req);
+      } else {
+        const confirmationUrl = this.getPublicConfirmationUrl(
+          dto.appUrl,
+          confirmationToken,
+        );
+        this.emailService.welcome(
+          jurisdictionName,
+          mapTo(User, newUser),
+          dto.appUrl,
+          confirmationUrl,
+        );
+      }
+
       // Partner user that is given access to an additional jurisdiction
     } else if (
       forPartners &&
@@ -875,5 +892,72 @@ export class UserService {
     }
 
     return false;
+  }
+
+  /**
+   *
+   * @param dto the incoming request with the email
+   * @returns a SuccessDTO always, and if the user exists it will send a code to the requester
+   */
+  async requestSingleUseCode(
+    dto: RequestSingleUseCode,
+    req: Request,
+  ): Promise<SuccessDTO> {
+    const user = await this.prisma.userAccounts.findFirst({
+      where: { email: dto.email },
+      include: {
+        jurisdictions: true,
+      },
+    });
+    if (!user) {
+      return { success: true };
+    }
+
+    const jurisdictionName = req?.headers?.jurisdictionname;
+    if (!jurisdictionName) {
+      throw new BadRequestException(
+        'jurisdictionname is missing from the request headers',
+      );
+    }
+
+    const juris = await this.prisma.jurisdictions.findFirst({
+      select: {
+        id: true,
+        allowSingleUseCodeLogin: true,
+      },
+      where: {
+        name: jurisdictionName as string,
+      },
+      orderBy: {
+        allowSingleUseCodeLogin: OrderByEnum.DESC,
+      },
+    });
+
+    if (!juris) {
+      throw new BadRequestException(
+        `Jurisidiction ${jurisdictionName} does not exists`,
+      );
+    }
+
+    if (!juris.allowSingleUseCodeLogin) {
+      throw new BadRequestException(
+        `Single use code login is not setup for ${jurisdictionName}`,
+      );
+    }
+
+    const singleUseCode = generateSingleUseCode();
+    await this.prisma.userAccounts.update({
+      data: {
+        singleUseCode,
+        singleUseCodeUpdatedAt: new Date(),
+      },
+      where: {
+        id: user.id,
+      },
+    });
+
+    await this.emailService.sendSingleUseCode(mapTo(User, user), singleUseCode);
+
+    return { success: true };
   }
 }
