+++ api/test/unit/services/auth.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { randomUUID } from 'crypto';
 import { sign } from 'jsonwebtoken';
-import { Response, Request } from 'express';
+import { Response } from 'express';
 import { ConfigService } from '@nestjs/config';
 import { MailService } from '@sendgrid/mail';
 import {
@@ -29,7 +29,7 @@ import { JurisdictionService } from '../../../src/services/jurisdiction.service'
 import { GoogleTranslateService } from '../../../src/services/google-translate.service';
 import { PermissionService } from '../../../src/services/permission.service';
 import { Jurisdiction } from '../../../src/dtos/jurisdictions/jurisdiction.dto';
-import { OrderByEnum } from '../../../src/enums/shared/order-by-enum';
+import { generateSingleUseCode } from '../../../src/utilities/generate-single-use-code';
 
 describe('Testing auth service', () => {
   let authService: AuthService;
@@ -609,12 +609,6 @@ describe('Testing auth service', () => {
     expect(prisma.userAccounts.update).not.toHaveBeenCalled();
   });
 
-  it('should generate mfa code', () => {
-    expect(authService.generateSingleUseCode().length).toEqual(
-      Number(process.env.MFA_CODE_LENGTH),
-    );
-  });
-
   it('should update password when correct token passed in', async () => {
     const id = randomUUID();
     const token = sign(
@@ -840,220 +834,5 @@ describe('Testing auth service', () => {
       'True',
       ACCESS_TOKEN_AVAILABLE_OPTIONS,
     );
-  });
-
-  it('should request single use code but user does not exist', async () => {
-    const id = randomUUID();
-    emailService.sendSingleUseCode = jest.fn();
-    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue(null);
-    prisma.userAccounts.update = jest.fn().mockResolvedValue({
-      id,
-    });
-
-    const res = await authService.requestSingleUseCode(
-      {
-        email: 'example@exygy.com',
-      },
-      { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
-    );
-
-    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
-      where: {
-        email: 'example@exygy.com',
-      },
-      include: {
-        jurisdictions: true,
-      },
-    });
-    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
-    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
-    expect(res).toEqual({
-      success: true,
-    });
-  });
-
-  it('should request single use code but jurisdiction does not exist', async () => {
-    const id = randomUUID();
-    emailService.sendSingleUseCode = jest.fn();
-    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
-      id,
-    });
-    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue(null);
-    prisma.userAccounts.update = jest.fn().mockResolvedValue({
-      id,
-    });
-
-    await expect(
-      async () =>
-        await authService.requestSingleUseCode(
-          {
-            email: 'example@exygy.com',
-          },
-          { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
-        ),
-    ).rejects.toThrowError('Jurisidiction juris 1 does not exists');
-
-    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
-      where: {
-        email: 'example@exygy.com',
-      },
-      include: {
-        jurisdictions: true,
-      },
-    });
-    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
-      select: {
-        id: true,
-        allowSingleUseCodeLogin: true,
-      },
-      where: {
-        name: 'juris 1',
-      },
-      orderBy: {
-        allowSingleUseCodeLogin: OrderByEnum.DESC,
-      },
-    });
-    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
-    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
-  });
-
-  it('should request single use code but jurisdiction disallows single use code login', async () => {
-    const id = randomUUID();
-    emailService.sendSingleUseCode = jest.fn();
-    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
-      id,
-    });
-    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
-      id: randomUUID(),
-      allowSingleUseCodeLogin: false,
-    });
-    prisma.userAccounts.update = jest.fn().mockResolvedValue({
-      id,
-    });
-
-    await expect(
-      async () =>
-        await authService.requestSingleUseCode(
-          {
-            email: 'example@exygy.com',
-          },
-          { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
-        ),
-    ).rejects.toThrowError('Single use code login is not setup for juris 1');
-
-    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
-      where: {
-        email: 'example@exygy.com',
-      },
-      include: {
-        jurisdictions: true,
-      },
-    });
-    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
-      select: {
-        id: true,
-        allowSingleUseCodeLogin: true,
-      },
-      where: {
-        name: 'juris 1',
-      },
-      orderBy: {
-        allowSingleUseCodeLogin: OrderByEnum.DESC,
-      },
-    });
-    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
-    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
-  });
-
-  it('should request single use code but jurisdictionname was not sent', async () => {
-    const id = randomUUID();
-    emailService.sendSingleUseCode = jest.fn();
-    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
-      id,
-    });
-    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
-      id,
-    });
-    prisma.userAccounts.update = jest.fn().mockResolvedValue({
-      id,
-    });
-
-    await expect(
-      async () =>
-        await authService.requestSingleUseCode(
-          {
-            email: 'example@exygy.com',
-          },
-          {} as unknown as Request,
-        ),
-    ).rejects.toThrowError(
-      'jurisdictionname is missing from the request headers',
-    );
-
-    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
-      where: {
-        email: 'example@exygy.com',
-      },
-      include: {
-        jurisdictions: true,
-      },
-    });
-    expect(prisma.jurisdictions.findFirst).not.toHaveBeenCalled();
-    expect(prisma.userAccounts.update).not.toHaveBeenCalled();
-    expect(emailService.sendSingleUseCode).not.toHaveBeenCalled();
-  });
-
-  it('should successfully request single use code', async () => {
-    const id = randomUUID();
-    emailService.sendSingleUseCode = jest.fn();
-    prisma.userAccounts.findFirst = jest.fn().mockResolvedValue({
-      id,
-    });
-    prisma.jurisdictions.findFirst = jest.fn().mockResolvedValue({
-      id,
-      allowSingleUseCodeLogin: true,
-    });
-    prisma.userAccounts.update = jest.fn().mockResolvedValue({
-      id,
-    });
-
-    const res = await authService.requestSingleUseCode(
-      {
-        email: 'example@exygy.com',
-      },
-      { headers: { jurisdictionname: 'juris 1' } } as unknown as Request,
-    );
-
-    expect(prisma.userAccounts.findFirst).toHaveBeenCalledWith({
-      where: {
-        email: 'example@exygy.com',
-      },
-      include: {
-        jurisdictions: true,
-      },
-    });
-    expect(prisma.jurisdictions.findFirst).toHaveBeenCalledWith({
-      select: {
-        id: true,
-        allowSingleUseCodeLogin: true,
-      },
-      where: {
-        name: 'juris 1',
-      },
-      orderBy: {
-        allowSingleUseCodeLogin: OrderByEnum.DESC,
-      },
-    });
-    expect(prisma.userAccounts.update).toHaveBeenCalledWith({
-      data: {
-        singleUseCode: expect.anything(),
-        singleUseCodeUpdatedAt: expect.anything(),
-      },
-      where: {
-        id,
-      },
-    });
-    expect(emailService.sendSingleUseCode).toHaveBeenCalled();
-    expect(res.success).toEqual(true);
   });
 });
