+++ shared-helpers/src/locales/general.json
@@ -20,6 +20,7 @@
   "account.signUpSaveTime.applyFaster": "Apply faster with saved application details",
   "account.signUpSaveTime.checkStatus": "Check on the status of an application at any time",
   "account.signUpSaveTime.resetPassword": "Simply reset your password if you forget it",
+  "account.signUpSaveTime.useACode": "Use a code to sign in without a password",
   "account.signUpSaveTime.subTitle": "Having an account will save you time by using saved application details, and allow you to check the status of an application at anytime.",
   "account.signUpSaveTime.title": "Sign up quickly and check application status at anytime",
   "account.settings.alerts.currentPassword": "Invalid current password. Please try again.",
@@ -211,15 +212,19 @@
   "application.household.preferredUnit.title": "What unit sizes are you interested in?",
   "application.household.primaryApplicant": "Primary Applicant",
   "application.name.dobHelper": "For example: 01 19 2000",
+  "application.name.dobHelper2": "This is collected to verify that you are at least 18 years old.",
   "application.name.emailPrivacy": "We will only use your email address to contact you about your application.",
   "application.name.firstName": "First Name",
+  "application.name.firstOrGivenName": "First or Given Name",
   "application.name.lastName": "Last Name",
+  "application.name.lastOrFamilyName": "Last or Family Name",
   "application.name.middleName": "Middle Name",
   "application.name.middleNameOptional": "Middle Name (optional)",
   "application.name.noEmailAddress": "I don't have an email address",
   "application.name.title": "What's your name?",
   "application.name.yourDateOfBirth": "Your Date of Birth",
   "application.name.yourEmailAddress": "Your Email Address",
+  "application.name.yourEmailAddressPwdlessHelper": "Enter your email address and we'll send you a code for a password-free sign in.",
   "application.name.yourName": "Your Name",
   "application.preferences.HOPWA.doNotConsider.label": "I don't want to be considered",
   "application.preferences.HOPWA.hopwa.description": "%{county} copy goes here…",
