+++ sites/public/src/pages/create-account.tsx
@@ -65,7 +65,20 @@ export default () => {
         listingIdRedirect
       )
 
-      setOpenModal(true)
+      if (process.env.showPwdless) {
+        const redirectUrl = router.query?.redirectUrl as string
+        const listingId = router.query?.listingId as string
+        let queryParams: { [key: string]: string } = { email: data.email, flowType: "create" }
+        if (redirectUrl) queryParams = { ...queryParams, redirectUrl }
+        if (listingId) queryParams = { ...queryParams, listingId }
+
+        await router.push({
+          pathname: "/verify",
+          query: queryParams,
+        })
+      } else {
+        setOpenModal(true)
+      }
     } catch (err) {
       const { status, data } = err.response || {}
       if (status === 400) {
@@ -108,7 +121,7 @@ export default () => {
                   </label>
 
                   <label className={styles["create-account-field"]} htmlFor="firstName">
-                    {t("application.name.firstName")}
+                    {t("application.name.firstOrGivenName")}
                   </label>
                   <Field
                     controlClassName={styles["create-account-input"]}
@@ -129,8 +142,6 @@ export default () => {
                   <Field
                     name="middleName"
                     register={register}
-                    label={t("application.name.middleNameOptional")}
-                    readerOnly
                     error={errors.middleName}
                     validation={{ maxLength: 64 }}
                     errorMessage={t("errors.maxLength")}
@@ -138,20 +149,18 @@ export default () => {
                   />
 
                   <label className={styles["create-account-field"]} htmlFor="lastName">
-                    {t("application.name.lastName")}
+                    {t("application.name.lastOrFamilyName")}
                   </label>
                   <Field
                     name="lastName"
                     validation={{ required: true, maxLength: 64 }}
                     error={errors.lastName}
                     register={register}
-                    label={t("application.name.lastName")}
                     errorMessage={
                       errors.lastName?.type === "maxLength"
                         ? t("errors.maxLength")
                         : t("errors.lastNameError")
                     }
-                    readerOnly
                     controlClassName={styles["create-account-input"]}
                   />
                 </CardSection>
@@ -170,7 +179,12 @@ export default () => {
                     errorMessage={t("errors.dateOfBirthErrorAge")}
                     label={t("application.name.yourDateOfBirth")}
                   />
-                  <p className={"field-sub-note"}>{t("application.name.dobHelper")}</p>
+                  <p className={`field-note ${styles["create-account-dob-age-helper"]}`}>
+                    {t("application.name.dobHelper2")}
+                  </p>
+                  <p className={`field-note ${styles["create-account-dob-example"]}`}>
+                    {t("application.name.dobHelper")}
+                  </p>
                 </CardSection>
 
                 <CardSection
@@ -187,6 +201,11 @@ export default () => {
                     register={register}
                     controlClassName={styles["create-account-input"]}
                     labelClassName={"text__caps-spaced"}
+                    note={
+                      process.env.showPwdless
+                        ? t("application.name.yourEmailAddressPwdlessHelper")
+                        : null
+                    }
                   />
                 </CardSection>
                 <CardSection
