+++ sites/public/styles/create-account.module.scss
@@ -24,3 +24,11 @@
 .create-account-label {
   margin-bottom: var(--bloom-s1);
 }
+
+.create-account-dob-age-helper {
+  margin-top: var(--seeds-s4);
+}
+
+.create-account-dob-example {
+  margin-top: var(--seeds-s2);
+}
